requirements.txt
pynput
pyperclip
google-generativeai
customtkinter
python-dotenv
pyinstaller
openai
anthropic
pystray
Pillow
